+++ app/Http/Controllers/PumpBrandsController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\PumpBrandUpdateRequest;
 use App\Http\Requests\PumpBrandStoreRequest;
 use App\Http\Resources\PumpBrandResource;
 use App\Models\Pumps\PumpBrand;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Redirect;
 use Inertia\Inertia;
@@ -17,9 +18,11 @@ class PumpBrandsController extends Controller
      * Show the form for creating a new resource.
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function create(): Response
     {
+        $this->authorize('brand_create');
         return Inertia::render('PumpBrands/Create');
     }
 
@@ -28,9 +31,11 @@ class PumpBrandsController extends Controller
      *
      * @param PumpBrandStoreRequest $request
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function store(PumpBrandStoreRequest $request): RedirectResponse
     {
+        $this->authorize('brand_create');
         PumpBrand::create($request->validated());
         return Redirect::route('pump_series.index');
     }
@@ -40,9 +45,11 @@ class PumpBrandsController extends Controller
      *
      * @param PumpBrand $pumpBrand
      * @return Response
+     * @throws AuthorizationException
      */
     public function edit(PumpBrand $pumpBrand): Response
     {
+        $this->authorize('brand_edit');
         return Inertia::render('PumpBrands/Edit', [
             'brand' => new PumpBrandResource($pumpBrand)
         ]);
@@ -54,9 +61,11 @@ class PumpBrandsController extends Controller
      * @param PumpBrandUpdateRequest $request
      * @param PumpBrand $pumpBrand
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function update(PumpBrandUpdateRequest $request, PumpBrand $pumpBrand): RedirectResponse
     {
+        $this->authorize('brand_edit');
         $pumpBrand->update($request->validated());
         return Redirect::route('pump_series.index');
     }
@@ -66,9 +75,11 @@ class PumpBrandsController extends Controller
      *
      * @param PumpBrand $pumpBrand
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function destroy(PumpBrand $pumpBrand): RedirectResponse
     {
+        $this->authorize('brand_delete');
         $pumpBrand->delete();
         return Redirect::route('pump_series.index');
     }
@@ -78,6 +89,7 @@ class PumpBrandsController extends Controller
      */
     public function restore($id): RedirectResponse
     {
+        $this->authorize('brand_restore');
         PumpBrand::withTrashed()->find($id)->restore();
         return Redirect::route('pump_series.index');
     }
